add role and permissions to Distribution Networks module

+++ Modules/DistributionNetwork/app/Http/Controllers/DistributionPointController.php
@@ -5,6 +5,7 @@ namespace Modules\DistributionNetwork\Http\Controllers;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 use Modules\DistributionNetwork\Http\Requests\DistributionPoint\StoreDistributionPointRequest;
 use Modules\DistributionNetwork\Http\Requests\DistributionPoint\UpdateDistributionPointRequest;
 use Modules\DistributionNetwork\Models\DistributionPoint;
@@ -21,7 +22,8 @@ class DistributionPointController extends Controller
     public static function middleware(): array
     {
         return [
-            new Middleware('role:Super Admin|Distribution Network Manager', only:['store','index','show', 'update', 'destroy']),
+            new Middleware('can:show_distribution_network_component', only: ['show']),
+            new Middleware('can:view_all_distribution_network_component', only: ['index']),
         ];
     }
 
@@ -33,7 +35,7 @@ class DistributionPointController extends Controller
      */
     public function __construct(DistributionPointService $pointService)
     {
-        $this->pointService =$pointService;
+        $this->pointService = $pointService;
     }
 
     /**
@@ -44,11 +46,12 @@ class DistributionPointController extends Controller
      */
     public function index(Request $request)
     {
-        $filters = $request->only(['status','type', 'distribution_network_id']);
+        $filters = $request->only(['status', 'type', 'distribution_network_id']);
         return $this->successResponse(
-                            'Operation succcessful'
-                            ,$this->pointService->getAllPoints($filters)
-                            ,200);
+            'Operation succcessful',
+            $this->pointService->getAllPoints($filters),
+            200
+        );
     }
 
     /**
@@ -62,9 +65,10 @@ class DistributionPointController extends Controller
     public function store(StoreDistributionPointRequest $request)
     {
         return $this->successResponse(
-                            'Created succcessful'
-                            ,$this->pointService->createPoint($request->validated())
-                            ,201);
+            'Created succcessful',
+            $this->pointService->createPoint($request->validated()),
+            201
+        );
     }
 
     /**
@@ -78,9 +82,10 @@ class DistributionPointController extends Controller
     public function show(DistributionPoint $point)
     {
         return $this->successResponse(
-                            'Operation succcessful'
-                            ,$this->pointService->showPoint($point)
-                            ,200);
+            'Operation succcessful',
+            $this->pointService->showPoint($point),
+            200
+        );
     }
 
     /**
@@ -96,8 +101,9 @@ class DistributionPointController extends Controller
     public function update(UpdateDistributionPointRequest $request, DistributionPoint $point)
     {
         return $this->successResponse(
-                            'Updated succcessful'
-                            ,$this->pointService->updatePoint($request->validated(),$point));
+            'Updated succcessful',
+            $this->pointService->updatePoint($request->validated(), $point)
+        );
     }
 
     /**
@@ -110,9 +116,13 @@ class DistributionPointController extends Controller
      */
     public function destroy(DistributionPoint $point)
     {
+        if (!Gate::allows('delete_distribution_network_component', $point)) {
+            return $this->errorResponse('Unauthorized', null, 403);
+        }
         $this->pointService->deletePoint($point);
         return $this->successResponse(
-                        'Deleted succcessful'
-                        ,null);
+            'Deleted succcessful',
+            null
+        );
     }
 }
